+++ go/go.mod
@@ -3,8 +3,6 @@ module go2fe
 go 1.13
 
 require (
-	github.com/gin-gonic/gin v1.5.0
-	github.com/howeyc/fsnotify v0.9.0 // indirect
 	github.com/mattn/go-colorable v0.1.6 // indirect
 	github.com/pilu/config v0.0.0-20131214182432-3eb99e6c0b9a // indirect
 	github.com/pilu/fresh v0.0.0-20190826141211-0fa698148017 // indirect
